Safari-609.1.11.1/LayoutTests/fast/forms/range/range-drag-expected.txt

Test for dragging operations of <input type=range>

Tests for range dragging from center.
readOnly=false, disabled=false
PASS input.value is "100"
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is lastChangeEventCounter + 1
readOnly=true
PASS input.value is "100"
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is lastChangeEventCounter + 1
disabled=true
PASS input.value is "50"
PASS inputEventCounter is lastInputEventCounter
PASS changeEventCounter is lastChangeEventCounter

Tests for range dragging from edge
readOnly=false, disabled=false
PASS input.value is "100"
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is lastChangeEventCounter + 1
readOnly=true
PASS input.value is "100"
PASS inputEventCounter is >= lastInputEventCounter + 1
PASS changeEventCounter is lastChangeEventCounter + 1
disabled=true
PASS input.value is "50"
PASS inputEventCounter is lastInputEventCounter
PASS changeEventCounter is lastChangeEventCounter

PASS successfullyParsed is true

TEST COMPLETE
